build working with WCv2
