+++ babel.config.json
@@ -1,0 +1,3 @@
+{
+  "plugins": ["@babel/plugin-transform-optional-chaining"]
+}
