+++ package.json
@@ -47,6 +47,7 @@
     "@terminal-fi/swappa": "^1.4.4",
     "@ubeswap/default-token-list": "^4.0.10",
     "@walletconnect/types": "^2.4.7",
+    "@walletconnect/utils": "^2.4.7",
     "@walletconnect/web3wallet": "^1.4.0",
     "axios": "^0.21.1",
     "better-sqlite3": "^7.1.2",
@@ -61,7 +62,8 @@
     "savingscelo-with-ube": "^1.5.0",
     "source-map-support": "^0.5.16",
     "tslib": "^2.2.0",
-    "wcv1/client": "npm:@walletconnect/client@^1.6.5",
+    "wcv1/client": "npm:@walletconnect/client@^1.8.0",
+    "wcv1/types": "npm:@walletconnect/types@^1.8.0",
     "web3": "1.3.6"
   },
   "resolutions": {
@@ -72,6 +74,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
+    "@babel/plugin-transform-optional-chaining": "^7.22.5",
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.12.10",
     "@celo/explorer": "^1.2.4",
@@ -88,6 +91,7 @@
     "@types/tmp": "^0.2.0",
     "@typescript-eslint/eslint-plugin": "^4.14.2",
     "@typescript-eslint/parser": "^4.14.2",
+    "babel-loader": "^8.0.0",
     "babel-plugin-transform-imports": "^2.0.0",
     "css-loader": "^5.0.1",
     "electron": "^11.4.6",
